% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_tensor.genlight}
\alias{as_tensor.genlight}
\title{Convert a \code{genlight} object from the \code{adegenet} package to a \code{torch_tensor}}
\usage{
\method{as_tensor}{genlight}(
  x,
  dtype = torch::torch_float(),
  device = NULL,
  requires_grad = FALSE,
  NA.method = c("asis", "mean", "zero"),
  by_pop = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{genlight} object to convert}

\item{dtype}{a \link[torch]{torch_dtype} instance}

\item{device}{a device creted with \code{\link[torch:torch_device]{torch_device()}}}

\item{requires_grad}{if autograd should record operations on the returned tensor.}

\item{NA.method}{a method to replace NA; asis: leave NAs as is; mean: replace by the mean allele frequencies; zero: replace by zero}

\item{by_pop}{Should the object be split by population?}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A \code{torch_tensor}
}
\description{
Convert a \code{genlight} object from the \code{adegenet} package to a \code{torch_tensor}
}
