% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_tensor}
\alias{as_tensor}
\title{Convert an object with genomic data to a \code{torch_tensor}}
\usage{
as_tensor(
  x,
  dtype = torch::torch_float(),
  device = NULL,
  requires_grad = FALSE,
  NA.method = c("asis", "mean", "zero"),
  ...
)
}
\arguments{
\item{x}{Object to be converted}

\item{dtype}{a \link[torch]{torch_dtype} instance}

\item{device}{a device creted with \code{\link[torch:torch_device]{torch_device()}}}

\item{requires_grad}{if autograd should record operations on the returned tensor.}

\item{NA.method}{a method to replace NA; asis: leave NAs as is; mean: replace by the mean allele frequencies; zero: replace by zero}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A \code{torch_tensor}
}
\description{
Convert an object with genomic data to a \code{torch_tensor}
}
